+++ NEWS.md
@@ -5,6 +5,8 @@
 * Added helper `cat_cli()`.
 * fix: default `dicts_suggest = NULL` for `dictionary_get_item()` and `dictionary_retrieve_item()` for backward compatibility.
 * fix a bug in `dictionary_sugar_inc_get`
+* Functions `messagef()`, `warningf()` and `stopf()` now have a `class` argument and also add
+  the additional class `mlr3message`, `mlr3warning` and `mlr3error`, respectively.
 
 # mlr3misc 0.16.0